refactor: make questionpy_sdk._package private
Packages should only be built using the CLI for now.
