+++ questionpy_sdk/_package/__init__.py
@@ -1,8 +1,8 @@
 #  This file is part of the QuestionPy SDK. (https://questionpy.org)
 #  The QuestionPy SDK is free software released under terms of the MIT license. See LICENSE.md.
 #  (c) Technische Universität Berlin, innoCampus <info@isis.tu-berlin.de>
-from questionpy_sdk.package._builder import build_qpy_package
-from questionpy_sdk.package._targets import BuildTarget, DirBuildTarget, ZipBuildTarget
+from questionpy_sdk._package._builder import build_qpy_package
+from questionpy_sdk._package._targets import BuildTarget, DirBuildTarget, ZipBuildTarget
 
 __all__ = [
     "BuildTarget",
