+++ tests/questionpy_sdk/commands/test_package.py
@@ -14,10 +14,10 @@ from click.testing import CliRunner
 
 from questionpy_common.constants import DIST_DIR, MANIFEST_FILENAME
 from questionpy_common.manifest import Bcp47LanguageTag
+from questionpy_sdk._package._helper import create_normalized_filename
 from questionpy_sdk.commands.package import package
 from questionpy_sdk.constants import PACKAGE_CONFIG_FILENAME
 from questionpy_sdk.models import PackageConfig
-from questionpy_sdk.package._helper import create_normalized_filename
 from questionpy_sdk.resources import EXAMPLE_PACKAGE
